+++ CHANGELOG.md
@@ -1,10 +1,14 @@
 # Change Log
 
-All notable changes to the "defect-heatmap" extension will be documented in this file.
+All notable changes to the "activity-heatmap" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
 ## [Unreleased]
+
+### Changed
+
+- Name to "Activity Heatmap"
 
 ## [1.3.2] - 2020-10-10